styled Primitive and moved InfoWindow to be a component
