+++ src/frontend/Pages/Map/components/InfoWindow.ts
