+++ src/frontend/Pages/Map/elements/abstracts/Primative.ts
@@ -1,9 +1,15 @@
-
+/**
+ * Primitive parent abstract class
+ * @abstract
+ */
 abstract class Primative {
+    protected readonly id: string;
+    protected readonly map: google.maps.Map;
+    
     /**
-     * Constructor for the Primative class
-     * @param id   id of the primative object
-     * @param map   map of the primative object
+     * Constructor for Primative class
+     * @param id id of primative object
+     * @param map map of primative object
      */
     constructor (id: string, map: google.maps.Map) {
         this.id = id;
@@ -11,16 +17,17 @@ abstract class Primative {
     }
 
     /**
-     * Gets the id
+     * Gets id
+     * @returns string id
      */
-    public getId() : string { return this.id; }
+    public getId() : string { 
+        return this.id; 
+    }
+
     /**
      * Tells if the primative is visible
      */
     abstract isVisible() : boolean;
-
-    protected readonly id: string;
-    protected readonly map: google.maps.Map;
 }
 
 export default Primative;
